+++ HISTORY.md
@@ -5,6 +5,7 @@
 ### 2021-06-15
 * extensions: tweaked world-map primitives
 * updated maps library
+* extensions: new naming convention
 
 ### 2021-06-14
 * new dev version